第一大块

+++ MVC-8/engage_resume.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -8,6 +9,7 @@ namespace MVC_8
 {
    public class engage_resume
     {
+        
         public int Id { get; set; }
         public string human_name { get; set; }
         public string engage_type { get; set; }
@@ -48,7 +50,7 @@ namespace MVC_8
         public string checker { get; set; }
         public DateTime check_time { get; set; }
         public int interview_status { get; set; }
-        public float total_points { get; set; }
+        public int total_points { get; set; }
         public int test_amount { get; set; }
         public string test_checker { get; set; }
         public DateTime test_check_time { get; set; }
